more tests

+++ SimpleWarehouseAPI/SimpleWarehouseAPI/Services/ProductService.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using SimpleWarehouseAPI.Models;
+using SimpleWarehouseAPI.Validator;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
@@ -29,23 +30,36 @@ namespace SimpleWarehouseAPI.Services
         
         public async Task<Product> GetProduct(Guid id)
         {
-            return await _context.Products.FirstOrDefaultAsync(x_ => x_.Id == id);
+            if (ProductValidator.IsIdOk(id))
+            {
+                return await _context.Products.FirstOrDefaultAsync(x_ => x_.Id == id);
+            }
+            else
+            {
+                return null;
+            }
         }
 
         public async Task<Guid?> AddProduct(ProductCreateInputModel model)
         {
-            Product product = new Product(model);
-            _context.Products.Add(product);
-            await _context.SaveChangesAsync();
-            return product.Id;
+            if (ProductValidator.ValidateData(model, out _))
+            {
+                Product product = new Product(model);
+                _context.Products.Add(product);
+                await _context.SaveChangesAsync();
+                return product.Id;
+            }
+            else
+            {
+                return null;
+            }
         }
         
         public async Task UpdateProduct(ProductUpdateInputModel model)
         {
-            Product foundedProduct = await _context.Products.FirstOrDefaultAsync(x_ => x_.Id == model.Id);
-
-            if (foundedProduct != null)
+            if (ProductValidator.ValidateData(model, _context, out _))
             {
+                Product foundedProduct = await _context.Products.FirstOrDefaultAsync(x_ => x_.Id == model.Id);
                 foundedProduct.Name = model.Name;
                 foundedProduct.Price = model.Price;
                 await _context.SaveChangesAsync();
@@ -54,13 +68,16 @@ namespace SimpleWarehouseAPI.Services
 
         public async Task DeleteProduct(Guid id)
         {
-            Product foundedProduct = await _context.Products.FirstOrDefaultAsync(x_ => x_.Id == id);
-
-            if (foundedProduct != null)
+            if (ProductValidator.IsIdOk(id))
             {
-                _context.Products.Remove(foundedProduct);
-                await _context.SaveChangesAsync();
-            }
+                Product foundedProduct = await _context.Products.FirstOrDefaultAsync(x_ => x_.Id == id);
+
+                if (foundedProduct != null)
+                {
+                    _context.Products.Remove(foundedProduct);
+                    await _context.SaveChangesAsync();
+                }
+            }       
         }
     }
 }
